Put the DataContext global in its own file to help HMR
